chord: add messaging types for basic chord rpcs
The implementation of the chord algorithm requires a number of different
remote procedure calls. This commit adds MessageTypes for these calls to
the MessageType enum and adds an implementation of the find successor
rpc.

Comms coder is updated to swap endianness for arbitrary sized types,
which is needed for encoding NodeId's