chord: fix the implementation of node ID addition
Node ID addition has been fixed so that it a) gives the correct results
and b) uses uint32_t integers for the addition.

To implement this fix, it was neccessary to allow the byte ordering of
the node IDs to be the same as the native byte ordering of the system.

The now unneeded field of m_id2 has been removed, this is not needed
anymore as this was an experimental field that had been added to support
a previously unsuccessful attempt at implementing the addition.

A constructor has been added that allows the Node ID to be created from
a std::string reprenting the big endian hex value of the node ID. This
can include dashes for readability, example:

  NodeId nodeId{ "a2fe349b-08cbe221-fd3ef3a2-7d76ada2-991bba1e" };
  NodeId nodeId{ "A2FE349B-08CBE221-FD3EF3A2-7D76ADA2-991BBA1E" };