comms: make Message::encode const
This allows messages to be passed to send functions and methods by const
reference. (send functions require the message to be encoded to an octet
string)

This change required changing the signature of a number of methods and
functions that are part of the call tree for implementations of encode
methods, as well as the all of the current concrete implementions of the
encode method (including one in the chord namespace).

The toEncode parameter passed to singleValueEncode has been changed to
be const.