chord: add ConnectionManager to simplify external comms
The ConnectionManager will hold the TcpServer establishing inbound
connections, and all of the TcpClients for managing outbound
connections.

The ConnectionManager class is used send all messages on to other nodes.
Because of this, the methods sendFindSuccessorRequest and
sendFindPredecessorRequest have been removed.

As the ConnectionManager contains a TcpServer instance, the TcpServer
owned by the ChordNode has been removed. Messages send to the node are
received by the TcpServer in the ConnectionManager and then sent on to
the ChordNode object via a callback.