chord: implement handling of received find successor messages
For the find successor algorithm to work, each chord node will need to
be able to process the find successor requests and responses.

This commit changes the previously incomplete processReceivedMessage to
handleReceivedMessage. The parameter that is passed to this method has
also been changed from a Message to an EncodedMessage.

The method currently ignores the version of the message and extracts the
message type, which it uses to properly decode the message and trigger
the appropriate processing.

The thread pool has been removed as has the async method findSuccessor,
due to the pending list these are no longer needed.

The doFindSuccessor method now returns void, instead of previously
returning NodeId. Since this method is now only going to trigger remote
calls, it should not be returing anything. Additionally, the parameter
for doFindSuccessor has been changed to a reference to a
FindSuccessorMessage, allowing all of the information from the received
message to be passed into the method.

In addition, some old unused and unimplemented methods have been
removed.