chord: add request id to messages requiring responses
If a message send to a node requires a response, then the sending node
needs to have a way of tracking when the response is received.

To do this, a request ID is added to the message being sent. The sending
node then saves the MessageType of the response that it expects to
receive and the node ID of the node that it expects to receive the
response from.

The sending node then saves this information in an unordered_map using
the request ID as a key.

If node A is sending a message that requires a response to node B then when
node B receives the message, it adds the request ID to the response
message and sends that response message back to node A.

Node A receives the response and extracts the request ID and uses that
to ID to check its unordered_map of responses it is waiting for.