chord: implement handling of find successor responses
Chord nodes need to be abole to handle receiving a find successor
response message. There are two possible ways such a message could be
handled:

1) if the receiving node is the one that started the procedure then the
   caller should be informed that the result has been received.

2) if the receiving node is just one of the nodes in the call chain then
   it needs to forward the message on to the node from which it received
   the original request.

Both of these cases are handled. When a node receives a find successor
response, it checks if m_hasChain is true in the pending structure, if
this is true then the message is repackaged with the current nodes ID as
the source ID and the message is send onwards.

If m_hasChain is false, then this must be the node that started the
procedure. Handling of this case has been implemented using a future and
a promise, and is currently only implemented in the join method (which
is the only place in which findSuccessor is currently used).

An unordered map of find successor promises has been added to the
ChordNode. When the find successor procedure is started a promise is
created and stored in the map with the key being the request ID.

When the find successor response is received at this node the received
ID can be given to the promise to make it available to whoever owns the
associated future.