chord: create ConnectionManager_I interface class
This interface has been created to allow the behaviour of the
ConnectionManager to be mocked in test code.

The ConnectionManager is now an implementation of this abstract
interface.

ChordNode now stores the ConnectionManager as a
std::unique_ptr<ConnectionManager_I>. The ChordNode constructor
accepts a factory lambda that creates a ConnectionManager in the
form of a std::unique_ptr<ConnectionManager_I>.

The use of the lambda due to the ConnectionManager requiring the ip
address and port to be used by the TCP server. If the ConnectionManager
where created outside of the class it is possible that a different ip
address and port could be given the ConnectionManager and the
ChordNode. This approach eliminates that problem as the factory lambda
creates the ConnectionManager from the same ip and port as are provided
to the ChordNode.

In ChordNodeTests a MockConnectionManager has been defined. In tests
that use ChordNodes, factory lambdas are created and provided to those
ChordNodes to create MockConnectionMangers that are linked to a
simulated network through a simulated node.