chord: fix decodeing of message type when handing receive message
The message type was being extracted directly from the m_message octet
string of the encoded message, and was not taking the endianness of the
system into account.

Messages are always encoded in big endian. The endianness must be
converted to system endianness.

The handling method now uses decodeSingleValue from Comms to ensure that
the value is extracted from the message correctly.