comms: add an EncodedMessage constructor for octet string
The new constructor allows an EncodedMessage to be constructed from an
existing encoded message in an array.

This is needed because when a message is received from the wire this
is the form that it will be in. This constructor will allow the entire
message to be dumped into an EncodedMessage object for further
processing.