chord: store on receive callback in MockConnectionManager
MockConnectionManager has a field for storing the on receive callback.
However, this field was not being used.

Now, when the callback is registered, it is stored in the object. The
object then passes the this pointer to the lambda that is send on the
simulated node. The lambda then calls the stored callback.