chord: refactor node join test and mark as tdd
The join test has been refactored so that only one connection factory
lambda is needed, this simplifies the code and also fixes a bug that was
caused by a copy of the simulated node when it was being returned from
addNode.

The test is still failing, and is going to require quite a lot of work
to fix. I have marked it as tdd for now so that it does trip up the build
pipeline.