chord: refactor finger table
The FingerTable class has been removed in favour of a struct and a
single function to intialise the finger table. This may be changed back
to a class at some point if complexity requires, however, the struct and
function are fine for now as they are.

Also, the FingerTableInterval struct has been removed in favour of just
storing the start and end points in the FingerTableEntry struct
directly.

IP and port have been removed from finger table entry. These are not
needed anymore as messages will be send by node ID, the code at this
level should not know the ip of the node that it wants to talk to.

Finger table tests have been removed, they will be added back if/when
the finger table starts doing something more interesting.