chord/netsim: add ip get method to connection manager
This get method is required so that the ChordNode can know its own IP
address, which it needs to know in order to send that address in a join
message when joining the network.

The MockConnectionManager does not store the IP address, and so the same
get method has been added to simulated node so that the mock manager can
get the ip from the simulated node.