chord: fix implementation of doFindSuccessor
The implementation of find successor was incorrect and hass been
improved.

If the ID is in the interval between the current node and the current
nodes successor, then the successor of the current node is returned.
Else, find the closest preceeding node and forward the find successor
response on to that node.

In addition, some basic log print statements have been added.