chord: change generation of node id when chord node is constructed
Everywhere else uses the uint32_t version of the IP address to create
the node ID. However, the ChordNode constructor was using a std::string
of the dotted decimal notation IP to generate its ip. This meant that
the ChordNode and the ConnectionManager have different IDs for each
node. This causes problems when ConnectionManager is asked to send a
message by node ID, as different parts of the system were using
different node IDs to refer to the same node.

The constructor has been changed so that it uses the uint32_t version of
the IP.

The method that ChordNode was using to generate the node ID from a
string has been removed.