chord: make predecessor node optional
The chord algorithm needs the predecessor value to be nil when the node
is created. The predecessor ID has been wrapped in a std::optional in
order to support this behaviour.

The get neighbours response has been updated to store an additional bool
field that represents if the nodes predecessor node was defined when the
message was sent. This change required a fix to CommsCoder to allow for
the encoding of a bool value.