chord: add lock free task queue to ChordNode
All async tasks (i.e., those requiring rpcs) are sent to this task
queue.

A work thread has been added to the ChordNode on which the work sent to
the queue will be processed.

m_predecessor no longer uses a std::optional.

Futures for join, findSuccessor and getNeighbours are now stored in the
ChordNode.